chore(Accordion): fix build error: Initializers are not allowed in ambient contexts.
