+++ packages/dnb-eufemia/src/components/accordion/Accordion.js
@@ -7,7 +7,6 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import classnames from 'classnames'
 import {
-  warn,
   isTrue,
   makeUniqueId,
   findElementInChildren,
@@ -27,26 +26,7 @@ import {
   accordionPropTypes,
   accordionDefaultProps,
 } from './AccordionPropTypes'
-
-class AccordionStore {
-  constructor(id) {
-    this._id = id
-    this._instances = []
-  }
-  onChange({ id }) {
-    this._instances.forEach((inst) => {
-      if (inst._id !== id) {
-        inst.close()
-      }
-    })
-  }
-  addInstance(instance) {
-    this._instances.push(instance)
-  }
-  removeInstance(instance) {
-    this._instances = this._instances.filter((inst) => inst !== instance)
-  }
-}
+import { AccordionStore, Store, rememberWarning } from './AccordionStore'
 
 export default class Accordion extends React.PureComponent {
   static contextType = AccordionProviderContext
@@ -452,104 +432,4 @@ Accordion.Group.Store = (group, id = null) => {
 }
 Accordion.Store = (id) => {
   return new Store({ id })
-}
-
-function rememberWarning(type = 'accordion') {
-  warn(`Missing "id" prop the ${type}! "remember_state" is enabled.`)
-}
-
-class Store {
-  constructor({ id, group }) {
-    this.id = id
-    this.group = group
-    return this
-  }
-
-  storeId(id = this.id) {
-    if (this.group) {
-      // Skip using the random ID
-      if (this.group[0] === '#') {
-        return null
-      }
-      id = this.group
-    }
-    return `dnb-accordion-${id}`
-  }
-
-  saveState(expanded, id = this.id, opts = {}) {
-    if (id) {
-      try {
-        const store = this.getData() || {}
-
-        if (this.group) {
-          if (expanded) {
-            store.id = id
-          } else if (opts && opts.force) {
-            store.id = null
-          }
-        } else {
-          store.expanded = expanded
-        }
-
-        const storeId = this.storeId(id)
-        if (storeId) {
-          window.localStorage.setItem(storeId, JSON.stringify(store))
-        }
-      } catch (e) {
-        //
-      }
-    } else {
-      rememberWarning()
-    }
-  }
-
-  getData(id = this.id) {
-    const storeId = this.storeId(id)
-
-    if (storeId) {
-      try {
-        if (
-          Object.prototype.hasOwnProperty.call(
-            window.localStorage,
-            storeId
-          )
-        ) {
-          return JSON.parse(window.localStorage.getItem(storeId))
-        }
-      } catch (e) {
-        //
-      }
-    }
-
-    return null
-  }
-
-  getState(id = this.id) {
-    let state = null
-
-    const store = this.getData(id)
-
-    if (store) {
-      if (typeof store.id !== 'undefined') {
-        state = id === store.id
-      } else if (store.expanded !== 'undefined') {
-        state = isTrue(store.expanded)
-      }
-    }
-
-    return state
-  }
-
-  flush(id = this.id) {
-    if (id) {
-      try {
-        const storeId = this.storeId(id)
-        if (storeId) {
-          window.localStorage.setItem(storeId, null)
-        }
-      } catch (e) {
-        //
-      }
-    }
-  }
 }
